refactor(moodle): pre-PR cleanup of feature/moodle
A two-track audit of the branch (dead code / refactor candidates +
upstream pattern alignment) surfaced both real technical debt and
several stylistic divergences that would draw review comments. This
commit folds the fixes into three phases.

Phase 1 -- PR blockers:
- MoodleClient now constructs a persistent httpx.AsyncClient with a
  30s default timeout (was instantiating one per call). Adds aclose()
  for graceful teardown.
- Drop the empty /lti/jwks endpoint that returned {"keys": []} -- it
  silently fails any platform attempting JWK validation against us.
- Drop unused jwcrypto direct dependency (still pulled transitively
  via python-keycloak; saves install weight).
- Fix AGENTS.md reference to deleted CLAUDE.md.

Phase 2 -- alignment + targeted refactor:
- moodle_tools_agent_factory now matches the AgentFactory protocol
  signature (kw-only args after agent_config, returns SoliplexAgent).
- Re-merge lti_templates.py back into views/lti.py to match how every
  other view module in the project keeps inline HTML.
- Move tests/unit/test_moodle_*.py into tests/unit/moodle/ to match
  the agui/, config/, skills/ subdirectory layout.
- Replace 102 try/except (MoodleAPIError, httpx.HTTPError) blocks
  with a single @_moodle_tool decorator (-200 LOC in skills.py).
- Introduce _preview() helper for the standard preview-JSON shape.
- Promote _resolve_rule_id to module level (was nested closure).
- Hoist _strip_html and _li_texts -> _parse_li_texts to module level.
- Switch from raw haiku.skills.SkillToolset to SoliplexSkillToolset
  for run-context state injection consistency with the rest of the
  project.
- Remove 7 dead public methods on MoodleClient (and their tests) that
  had zero callers anywhere in src/ or tests/.
- Rename tests/unit/_test_features.py -> _fixtures.py so its real
  role as a shared fixture module is clear (the _test_ prefix was
  fooling pytest into skipping it as a test file).
- Document the [tool.coverage.run].omit list in pyproject.toml.

Phase 3 -- polish (preview-leak fix):
- Drop the "instructions" field from preview JSON entirely. The
  string ("Present this to the user and ask for confirmation...")
  was being rendered verbatim into user-facing tables (smoke
  scenarios 29, 34, 44). The skill prompt's Case A / Case B already
  covers preview-vs-confirm handling, so the field was redundant.
- Strengthen _CONFIRM_INSTRUCTIONS with explicit "no preamble, no
  narration" output rules so chain-of-thought style summaries
  ("We identified...", "User wants to...") stop leaking before the
  preview table.

Verification:
- 4,342 unit tests pass; 100% coverage maintained.
- ruff check + format clean.
- Re-ran all 57 MOODLE.md smoke scenarios via REST against
  gpt-oss:latest: 47 PASS / 7 PARTIAL / 3 FAIL. None of the
  non-passes are caused by this refactor -- they are gpt-oss
  hallucination on empty tool results, sandbox state pollution from
  prior runs, or seed-data gaps (UTM/adv-comp report tables empty).
- Re-tested the three preview-leak scenarios (29, 34, 44) after
  Phase 3 changes; all now produce clean table-only previews.
