+++ tests/unit/conftest.py
@@ -2,7 +2,7 @@ import pathlib
 import tempfile
 from unittest import mock
 
-import _test_features as agui_features
+import _fixtures as agui_features
 import httpx
 import pytest
